btcapi/requirements.txt
falcon
simplejson
python-bitcoinrpc
gunicorn
pycrypto